Use npm@3.3.9 to manage npm dependencies of packages.
We were already using npm@3.1.2 to install the contents of
dev_bundle/{lib,server-lib}/node_modules, with care to restore the
original npm@1.4.28 before bundling. This commit uses npm@1.4.28 to
install the latest version of npm (3.3.9) over the original version.

This upgrade allowed significant simplifications to meteor-npm.js, most
notably that it is no longer necessary to install into a temporary package
directory and then rename the directory upon success, because npm 3 does
that itself (see the ALGORITHM section of `npm help install`).